Incorporate Pull Request suggestions
