+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "mcp-atlassian"
-version = "0.1.11"
+version = "0.1.10"
 description = "The Model Context Protocol (MCP) Atlassian integration is an open-source implementation that bridges Atlassian products (Jira and Confluence) with AI language models following Anthropic's MCP specification. This project enables secure, contextual AI interactions with Atlassian tools while maintaining data privacy and security. Key features include:"
 readme = "README.md"
 requires-python = ">=3.10"
@@ -11,8 +11,7 @@ dependencies = [
     "mcp>=1.0.0",
     "python-dotenv>=1.0.1",
     "markdownify>=0.11.6",
-    "markdown>=3.7.0",
-    "markdown-to-confluence>=0.3.0"
+    "markdown>=3.7.0"
 ]
 [[project.authors]]
 name = "sooperset"
